refactor: correct packages for test classes
